Refactor schema: split project.ts into component.ts and codeblock.ts modules
