+++ apps/server/schema/codeblock.ts
@@ -1,0 +1,49 @@
+import {
+  jsonb,
+  pgTable,
+  serial,
+  text,
+  timestamp,
+  uuid,
+} from "drizzle-orm/pg-core";
+import { InferSelectModel, relations } from "drizzle-orm";
+import { Page } from "./component";
+import { Project, Workspace } from "./user";
+
+// ++++++++++++++++++++++++++++++++++++++++++++++++ CodeBlock Tables +++++++++++++++++++++++++++++++++++++++++++++++++
+export const CodeBlock = pgTable("codeblock", {
+  id: uuid("id").defaultRandom().primaryKey(),
+  project: uuid("project_id"),
+  name: text("name").notNull(),
+  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow(),
+});
+
+export type CodeBlockInterface = InferSelectModel<typeof CodeBlock>;
+
+export const StepBlock = pgTable("stepblock", {
+  id: uuid("id").defaultRandom().primaryKey(),
+  codeblock: uuid("codeblock_id"),
+  name: text("name").notNull(),
+  code: text("code").notNull(),
+  language: text("language").notNull(),
+  output: text("output").notNull(),
+  stdout: text("stdout").notNull(),
+  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow(),
+});
+
+export type StepBlockInterface = InferSelectModel<typeof StepBlock>;
+
+export const codeBlockRelations = relations(CodeBlock, ({ one, many }) => ({
+  project: one(Project, {
+    fields: [CodeBlock.project],
+    references: [Project.id],
+  }),
+  stepBlocks: many(StepBlock),
+}));
+
+export const stepBlockRelations = relations(StepBlock, ({ one, many }) => ({
+  codeBlock: one(CodeBlock, {
+    fields: [StepBlock.codeblock],
+    references: [CodeBlock.id],
+  }),
+}));
