+++ apps/server/schema/index.ts
@@ -1,4 +1,5 @@
 // Barrel File for all Schema
 
 export * from "./user";
-export * from "./project";
+export * from "./component";
+export * from "./codeblock";
